// pcf-form/src/types/FormStep3.ts
export interface FormStep3Input {
  teamCoordinator: 'Sim' | 'Não' | 'Não sei';
  meetTheCoordinator: 'sim' | 'não';
  contactCoordinator: '0' | '1' | '2' | '3' | '4 ou mais';
  knowTheMultiplier: 'Sim' | 'Não';
  steeringCommittee: 'Sim' | 'Não' | 'Não sei';
  steeringCommitteeMeeting: '0' | '1' | '2' | '3' | '4 ou mais';
};